pyproject.toml
[tool.poetry]
name = "backend-technical-test-master"
version = "0.1.0"
description = ""
authors = ["kAldown <kaldownb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
psycopg2 = "^2.9.1"
PyYAML = "^5.4.1"
uritemplate = "^4.0.0"
coreapi = "^2.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
